Finish work to setup cell with viewmodel + fix bug where panel state was changing for separator creation
